+++ src/linter/citrine.cpp
@@ -4,9 +4,6 @@
 #include <regex>
 #include <algorithm>
 #include <iostream>
-#ifdef _WIN32
-#include <windows.h>
-#endif
 
 namespace citrine {